removed big test file and added remaining parts
